+++ src/rogallo/commands/main.py
@@ -100,4 +100,18 @@ class ClearCache(Command):
     BINDING_KEY = "shift+f5"
 
 
+##############################################################################
+class StripeLinks(Command):
+    """Toggle the striping of links in the document viewer"""
+
+    BINDING_KEY = "f8"
+
+
+##############################################################################
+class ToggleLinkNumbers(Command):
+    """Toggle the display of link numbers in the document viewer"""
+
+    BINDING_KEY = "shift+f8"
+
+
 ### main.py ends here