+++ src/rogallo/providers/main.py
@@ -29,8 +29,10 @@ from ..commands import (
     SearchHistory,
     SetHome,
     SetHomeToCurrentLocation,
+    StripeLinks,
     ToggleBookmarks,
     ToggleHistory,
+    ToggleLinkNumbers,
     ToggleView,
 )
 
@@ -64,8 +66,10 @@ class MainCommands(CommandsProvider):
         yield from self.maybe(SearchHistory)
         yield SetHome()
         yield from self.maybe(SetHomeToCurrentLocation)
+        yield StripeLinks()
         yield from self.maybe(ToggleBookmarks)
         yield from self.maybe(ToggleHistory)
+        yield ToggleLinkNumbers()
         yield from self.maybe(ToggleView)