+++ src/rogallo/widgets/viewer/widget.py
@@ -13,7 +13,7 @@ from gemtext import Gemtext, Line, Paragraph
 from textual import on
 from textual.app import ComposeResult
 from textual.containers import Vertical
-from textual.events import DescendantBlur, DescendantFocus
+from textual.events import DescendantBlur, DescendantFocus, Key
 from textual.getters import query_one
 from textual.reactive import var
 from textual.widgets import Static
@@ -22,7 +22,7 @@ from textual.widgets import Static
 # Local imports.
 from ...document import Document
 from .document_view import DocumentView
-from .gemtext_blocks import GemtextLink, GemtextWidget, get_block_widget
+from .gemtext_blocks import GemtextLink, get_block_widget
 from .status import ViewerStatus
 from .title import ViewerTitle
 
@@ -40,6 +40,14 @@ class Viewer(Vertical, can_focus=False):
         &.--has-content {
             visibility: visible;
         }
+
+        &.--stripe-links GemtextLink.--stripe {
+            background: $boost 200%;
+        }
+
+        &.--with-link-numbers GemtextLink #jump {
+            display: block;
+        }
     }
     """
 
@@ -47,6 +55,10 @@ class Viewer(Vertical, can_focus=False):
     """The details of the document to show in the viewer."""
     view_source: var[bool] = var(False)
     """Whether the viewer is showing the source of the document or not."""
+    with_link_numbers: var[bool] = var(False, toggle_class="--with-link-numbers")
+    """Whether the viewer is showing link numbers or not."""
+    stripe_links: var[bool] = var(False, toggle_class="--stripe-links")
+    """Whether the viewer is showing links with stripes or not."""
 
     _title = query_one(ViewerTitle)
     """The title widget."""
@@ -54,6 +66,9 @@ class Viewer(Vertical, can_focus=False):
     """The document view widget."""
     _status = query_one(ViewerStatus)
     """The status bar widget."""
+
+    _jump: var[int | None] = var(None)
+    """Keeps track of the jump progress."""
 
     def compose(self) -> ComposeResult:
         """Compose the viewer widget."""
@@ -89,26 +104,22 @@ class Viewer(Vertical, can_focus=False):
         self._title.location = self.document.location
         self._status.mime_type = self.document.mime_type or ""
         await self._view.remove_children()
-        blocks: list[Static] | list[GemtextWidget]
-        if not self.document.is_gemtext or self.view_source:
-            blocks = [
+        await self._view.mount_all(
+            [
                 Static(
                     self.document.content.replace(chr(27), "\N{SYMBOL FOR ESCAPE}"),
                     markup=False,
                 )
             ]
-        else:
-            for widget in (
-                blocks := [
-                    get_block_widget(line)
-                    for line in self._consolidate(
-                        Gemtext(self.document.content).content
-                    )
-                ]
-            ):
-                if isinstance(widget, GemtextLink):
-                    widget.normalise_uri(self.document.location)
-        await self._view.mount_all(blocks)
+            if not self.document.is_gemtext or self.view_source
+            else [
+                get_block_widget(line)
+                for line in self._consolidate(Gemtext(self.document.content).content)
+            ]
+        )
+        for jump_number, link in enumerate(self._view.query(GemtextLink)):
+            link.normalise_uri(self.document.location)
+            link.jump_number = jump_number + 1
         # This next bit of nonsense is because Textual fails to sort its
         # scrollbars out upon clearing down and remounting a new set of
         # children. So we have to force it to refresh and then scroll to the
@@ -121,6 +132,19 @@ class Viewer(Vertical, can_focus=False):
     def _watch_view_source(self) -> None:
         """Watch for changes to the view_source property and update the viewer."""
         self.mutate_reactive(Viewer.document)
+
+    def _watch_with_link_numbers(self) -> None:
+        """Watch for changes to the with_link_numbers property."""
+        self._jump = None
+
+    def _watch__jump(self) -> None:
+        """Watch for changes to the jump property and update the viewer."""
+        if self.with_link_numbers and self._jump is not None:
+            for link in self._view.query(GemtextLink):
+                if link.jump_number == self._jump:
+                    link.focus(scroll_visible=True)
+                    return
+            self._jump = self._jump % 10 if self._jump > 9 else None
 
     def take_control(self) -> None:
         """Take control of the UI."""
@@ -139,6 +163,17 @@ class Viewer(Vertical, can_focus=False):
         """Clear the status bar when a descendant widget is blurred."""
         if self.screen.focused and self not in self.screen.focused.ancestors:
             self._status.message = ""
+
+    @on(Key)
+    def _jumper(self, event: Key) -> None:
+        """Handle jump key presses."""
+        if not self.with_link_numbers:
+            return
+        if event.key.isdigit():
+            event.stop()
+            self._jump = (self._jump or 0) * 10 + int(event.key)
+        else:
+            self._jump = None
 
 
 ### widget.py ends here